+++ CHANGELOG.md
@@ -19,6 +19,8 @@
   - AppleALC 1.6.3 release
   - WhateverGreen 1.5.2 release
   - FeatureUnlock 1.0.3 release
+- Allow iGPU/dGPU switching in Windows
+  - Applicable to MacBook Pros with Intel iGPU and Nvidia/AMD dGPU
 
 ## 0.2.4